moving guess success test to appropriate file
