+++ tests/unit/resolver/test_issue_handler.py
@@ -1,8 +1,7 @@
 from unittest.mock import MagicMock, patch
 
 from openhands.core.config import LLMConfig
-from openhands.events.action.message import MessageAction
-from openhands.resolver.github_issue import GithubIssue, ReviewThread
+from openhands.resolver.github_issue import ReviewThread
 from openhands.resolver.issue_definitions import IssueHandler, PRHandler
 
 
@@ -45,55 +44,6 @@ def test_get_converted_issues_initializes_review_comments():
         assert issues[0].body == 'Test Body'
         assert issues[0].owner == 'test-owner'
         assert issues[0].repo == 'test-repo'
-
-
-def test_pr_handler_guess_success_with_thread_comments():
-    # Create a PR handler instance
-    llm_config = LLMConfig(model='test', api_key='test')
-    handler = PRHandler('test-owner', 'test-repo', 'test-token', llm_config)
-
-    # Create a mock issue with thread comments but no review comments
-    issue = GithubIssue(
-        owner='test-owner',
-        repo='test-repo',
-        number=1,
-        title='Test PR',
-        body='Test Body',
-        thread_comments=['First comment', 'Second comment'],
-        closing_issues=['Issue description'],
-        review_comments=None,
-        thread_ids=None,
-        head_branch='test-branch',
-    )
-
-    # Create mock history
-    history = [MessageAction(content='Fixed the issue by implementing X and Y')]
-
-    # Create mock LLM config
-    llm_config = LLMConfig(model='test-model', api_key='test-key')
-
-    # Mock the LLM response
-    mock_response = MagicMock()
-    mock_response.choices = [
-        MagicMock(
-            message=MagicMock(
-                content="""--- success
-true
-
---- explanation
-The changes successfully address the feedback."""
-            )
-        )
-    ]
-
-    # Test the guess_success method
-    with patch('litellm.completion', return_value=mock_response):
-        success, success_list, explanation = handler.guess_success(issue, history)
-
-        # Verify the results
-        assert success is True
-        assert success_list == [True]
-        assert 'successfully address' in explanation
 
 
 def test_pr_handler_get_converted_issues_with_comments():
@@ -177,87 +127,6 @@ def test_pr_handler_get_converted_issues_with_comments():
             assert prs[0].closing_issues == [
                 'This is additional context from an externally referenced issue.'
             ]
-
-
-def test_pr_handler_guess_success_only_review_comments():
-    # Create a PR handler instance
-    llm_config = LLMConfig(model='test', api_key='test')
-    handler = PRHandler('test-owner', 'test-repo', 'test-token', llm_config)
-
-    # Create a mock issue with only review comments
-    issue = GithubIssue(
-        owner='test-owner',
-        repo='test-repo',
-        number=1,
-        title='Test PR',
-        body='Test Body',
-        thread_comments=None,
-        closing_issues=['Issue description'],
-        review_comments=['Please fix the formatting', 'Add more tests'],
-        thread_ids=None,
-        head_branch='test-branch',
-    )
-
-    # Create mock history
-    history = [MessageAction(content='Fixed the formatting and added more tests')]
-
-    # Create mock LLM config
-    llm_config = LLMConfig(model='test-model', api_key='test-key')
-
-    # Mock the LLM response
-    mock_response = MagicMock()
-    mock_response.choices = [
-        MagicMock(
-            message=MagicMock(
-                content="""--- success
-true
-
---- explanation
-The changes successfully address the review comments."""
-            )
-        )
-    ]
-
-    # Test the guess_success method
-    with patch('litellm.completion', return_value=mock_response):
-        success, success_list, explanation = handler.guess_success(issue, history)
-
-        # Verify the results
-        assert success is True
-        assert success_list == [True]
-        assert 'successfully address' in explanation
-
-
-def test_pr_handler_guess_success_no_comments():
-    # Create a PR handler instance
-    llm_config = LLMConfig(model='test', api_key='test')
-    handler = PRHandler('test-owner', 'test-repo', 'test-token', llm_config)
-
-    # Create a mock issue with no comments
-    issue = GithubIssue(
-        owner='test-owner',
-        repo='test-repo',
-        number=1,
-        title='Test PR',
-        body='Test Body',
-        thread_comments=None,
-        closing_issues=['Issue description'],
-        review_comments=None,
-        thread_ids=None,
-        head_branch='test-branch',
-    )
-
-    # Create mock history
-    history = [MessageAction(content='Fixed the issue')]
-
-    # Create mock LLM config
-    llm_config = LLMConfig(model='test-model', api_key='test-key')
-
-    # Test that it returns appropriate message when no comments are present
-    success, success_list, explanation = handler.guess_success(issue, history)
-    assert success is False
-    assert success_list is None
-    assert explanation == 'No feedback was found to process'
 
 
 def test_get_issue_comments_with_specific_comment_id():
